fix sans notebook and make single get_detector_data
